# openapidocumentation.yaml
openapi: 3.0.0
info:
  title: 'GlintHub OpenAPI'
  version: 1.0.0
  description: 'Web Resources Specification for GlintHub'
servers:
- url: db.feup.up.pt
  description: Production server
...
paths:
...
  /login:
    get:
      operationId: R102
      summary: 'R102 : Login Form'
      description: 'Provide login form. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'
      responses:
        '200':
          description: 'Ok. Show log-in UI'
        '302':
          description: 'Redirect after verifying user credentials'
          headers:
            /home:
              description: 'You are already logged in!'
              schema:
                type: string

    post:
      operationId: R103
      summary: 'R103 : Login Action'
      description: 'Processes the login form submission. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'
      requestBody:
        required: true
        content:
          application/x-www-form-urllencoded:
            schema:
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                  - email
                  - password
      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          headers:
            /home:
              description: 'You have logged in successfully!'
              schema:
                type: string
            /login:
              description: 'The provided credentials do not match our records.'
              schema:
                type: string
...